backend/app/requirements.txt
fastapi==0.129.0
pydantic==2.12.5
pydantic_settings==2.13.0
python-dotenv==1.2.1
python_jose==3.5.0
Requests==2.32.5
SQLAlchemy==2.0.38
